+++ packages/paraglide/paraglide-js/src/cli/main.ts
@@ -1,6 +1,6 @@
 import { Command } from "commander"
-import { compileCommand } from "./commands/compile.js"
-import { initCommand } from "./commands/init.js"
+import { compileCommand } from "./commands/compile/command.js"
+import { initCommand } from "./commands/init/command.js"
 import { version } from "./state.js"
 
 export const cli = new Command()